style(folder structure): moved user controller to an inside folder
